+++ TCPApp/AppDelegate.swift
@@ -1,0 +1,30 @@
+//
+//  AppDelegate.swift
+//  TCPApp
+//
+//  Created by nghiatran on 24/4/26.
+//
+
+import Cocoa
+
+@main
+class AppDelegate: NSObject, NSApplicationDelegate {
+
+    
+
+
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        // Insert code here to initialize your application
+    }
+
+    func applicationWillTerminate(_ aNotification: Notification) {
+        // Insert code here to tear down your application
+    }
+
+    func applicationSupportsSecureRestorableState(_ app: NSApplication) -> Bool {
+        return true
+    }
+
+
+}
+
