+++ TCPApp/Assets.xcassets/Contents.json
