+++ TCPApp/Features/Inspector/PacketInspectorViewController.swift
@@ -1,0 +1,17 @@
+//
+//  PacketInspectorViewController.swift
+//  TCPMan
+//
+//  Created by nghiatran on 24/4/26.
+//
+
+import Cocoa
+
+class PacketInspectorViewController: NSViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do view setup here.
+    }
+    
+}
