+++ TCPApp/Features/PacketList/PacketListViewController.swift
@@ -1,0 +1,17 @@
+//
+//  PacketListViewController.swift
+//  TCPMan
+//
+//  Created by nghiatran on 24/4/26.
+//
+
+import Cocoa
+
+class PacketListViewController: NSViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do view setup here.
+    }
+    
+}
