+++ TCPApp/Features/Sidebar/SidebarViewController.swift
@@ -1,0 +1,17 @@
+//
+//  SidebarViewController.swift
+//  TCPMan
+//
+//  Created by nghiatran on 24/4/26.
+//
+
+import Cocoa
+
+class SidebarViewController: NSViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do view setup here.
+    }
+    
+}
