+++ TCPAppTests/TCPAppTests.swift
@@ -1,0 +1,19 @@
+//
+//  TCPAppTests.swift
+//  TCPAppTests
+//
+//  Created by nghiatran on 24/4/26.
+//
+
+import Testing
+@testable import TCPApp
+
+struct TCPAppTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+        // Swift Testing Documentation
+        // https://developer.apple.com/documentation/testing
+    }
+
+}
